Embed a live scanner audio player
Doesn't work on mobile at the moment, but that's ok.  The link
gracefully degrades to just sending you to the KOMO page.

I'm sure there are tons of web-based scanners out there, but the KOMO
one is a) the one I stumbled across today, b) seems reliable, and c) is
from a local news org so likely to persist.